Add push signature and qr
