+++ test/bin/Debug/test.exe.config
@@ -9,8 +9,8 @@
     <add key ="AUTHENTICATION_API_ACCESS_TOKEN" value ="b33ed51f7d87f29c4197389ca7688f8b182d49d4"/>
     <add key ="ADMIN_API_ACCESS_TOKEN" value ="a8f09bedadb87190576636bfe576ac8265969d87"/>
     
-    <add key ="INTERNAL_USERNAME" value ="internal_auth"/>
-    <add key ="INTERNAL_PASSWORD" value ="12345"/>
+    <add key ="INTERNAL_USERNAME" value ="spalomares"/>
+    <add key ="INTERNAL_PASSWORD" value ="Safewalk1"/>
     
     <add key ="LDAP_USERNAME" value ="test@local.com"/>
     <add key ="LDAP_PASSWORD" value ="12345"/>
@@ -25,6 +25,10 @@
     <add key ="USER_mobilePhone_new" value="+3812332131"/>
     <add key ="USER_email_new" value="g123321@gmail.com"/>
     <add key ="USER_password_new" value="abcde"/>
-
+    <!-- push signature -->
+    <add key ="hash" value="A160E4F805C51261541F0AD6BC618AE10BEB3A30786A099CE67DBEFD4F7F929F"/>
+    <add key ="data" value="All the data here will be signed. This request was generated from Safewalk API."/>
+    <add key ="title" value="Sign transaction"/>
+    <add key ="body" value="Push signature triggered from safewalk API"/>
   </appSettings>
 </configuration>
